Changed `class` to `model`
This was raised in #5.  This gets rid of `class`, which I agree is a good
thing.  But one lingering question is whether `model` is correct or more
appropriate.  I need to check the specification to see if it explicitly
specified the restricted class of `ExternalObject`.  I cannot imagine why
it would be a `model` though, since it doesn't have (nor can it have) equations.

Unfortunately, the Modelica server is currently down and I don't have a copy
of the specification handy.  So resolving this completely will have to wait
until Monday.