Referencing and outline
I added the headings for some of the remainder of chapter 1.  Also, I
think I figured out the issues with referencing.  It was useful to watch
the warnings coming out of Sphinx to understand the issue.  This makes
me think I should suppress all the junk output.  There is an option to
Sphinx to do that, I should probably add it to the Makefile to make sure
I address any real warnings.