Greatly improved structure
It turns out that the 'dirhtml' HTML builder appears to pretty much builds
one web page per file.  So I've broken things up into many more files.  This
means that each individual web page is much more succint.  This is a great
improvement in structure.  Previously, the amount of content on a page
was overwhelming.  This is much closer to the 'bite-sized chunks' that
I wanted to achieve.  I should carry this approach forward into each
chapter.