Added some TODO items to the generate_results script
I don't have Jinjify installed on this computer.  I really need to push it up onto
GitHub and/or PyPI and really define it as a standalone tool.

Once I have access in Jinjify then I should convert this whole generate_results mess
to be a descriptive file that generates all the plot scripts AND a Makefile so that
building everything can be done efficiently.