Working on refactoring the ABCD and LTI stuff
I need to refactor this to make it easier to explain.  In addition, there seem to be some
issues with OpenModelia figuring out some fo the parameters.  So I need to make sure that
this behavior is portable.

My plan is to refactor the examples so that some use ABCD and then others extend from
LTI (the final version, which should have a clean name).