Decided to rework these models
The issue was that starting from the most general to the least general was confusing.  Also, I tried
to demonstrate some 'clever' stuff that isn't supported by OpenModelica as well.  So I just cleaned
things up and decided to stick to the basics and present the material in a more coherent way.