No need for `:nowrap:` and explicit `align` environment
According to http://sphinx-doc.org/latest/ext/math.html we
do not need to use the `:nowrap:`  argument hance making the
code a bit cleaner. In case we need numbers a simple `:label: foo`
will generate a equation number automatically.

I tested this for both the PDF and the HTML output.