Tweaking the sensor comparison stuff
A big issue was that I failed to specify a value for the period parameter in my setpoint
component.  This was causing OM some problems (rightfully so).  I fixed that and then played
with some gains a bit.  Things look OK, but there are some differences in the solutions
between Dymola and OM that I need to investigate...