Working on my own specific CSS style for the book
I started playing around with this, and I like the degree of control and the simplicity of
just writing the CSS myself.  Of course, I'm stealing stuff left and right to put in here,
but this also keeps things pretty light.

Note that because of the way Sphinx creates structure and classes, I end up having to
write stuff explicitly anyway.  I could be clever and write some Javascript or something
to walk the tree and find everything of class X and add a class Y or use SASS and something
like Bootstrap and use CSS inheritance in some clever way.  But I just find this much
simpler.