Improved TOC by adding parents
This provides more context (@dietmarw was concerned about this, not sure if it
actually addresses his concerns).

It would be nice if I could get siblings of parent objects (need to check if
Sphinx provides that).  If so, I could effectively include a chapter list in
the TOC and show the structure for the current chapter/section in detail.