Adding po files to version control
It might seem odd to add the po files to version control since they are in some
sense generated during the compile process.  But since people will be editing these
files and because I need to easily track differences as the English version of the
book evolves, I think we need to version control the files.

I'm going to stop short of version controlling the pot files for now.  But I might
choose to do that as well.  I need to try out the translation process first to see
if that makes sense.