Updates to Spanish translation files
I'm not 100% sure why all of these updated.  I suspect because of changes in the book
text or something.  In any case, things seem to be in order with respect to the
dedication text (as far as **I** can tell).  I had to make a few minor changes to
whitespace.