Fixes #181 by making TODO list appear ONLY if condition is true
We can simply reuse the `todo_include_todos` condition.
However there is a little quirk to it when trying to hide
the TODO entry from the TOC. The conditional entry can only
be of lower level than the preceeding one. So that's why I
changed the level and put it under the "Indices and tables"
section.