Added a basic Go webhook handler
I'm pushing this out because the Docker build actually requires getting the Go
application from GitHub when building the image.  Perhaps I should find a way
to get it to use only the local version to avoid lots of extra commits.  But I'm
just testing the build process (not the application) so I think just getting it
bootstrapped should be sufficient.