Creating various stages images
The point of this is that images that require source code (the MBE and PB images)
shouldn't used cached images.  But I want to use cached images for the other two
because that will save a lot of build time.  Doing it this way makes it relatively
easy to update the images by using caching when I should and not when I shouldn't.

The last image (PB, short for pubbook) runs the 'pubbook' command so the container
is all set to be run as an application container.  Now I just need to make sure
the POST handler does what it needs to do.  That I can test locally without
really having to build the container until I need to deploy it.