Fix the hierarchy for the Front Matter in LaTeX
This is an ugly hack but necessary for the LaTeX generation.
Basically the problem is that in order to get the Chapters
as \chapters generated there needs to be a ### heading present
before the *** Front Matter ***. If not Sphinx will
generate the chapters as \part in LaTeX. I tried conditional
`.. only:: latex` syntax but that does not work with top level
section (needs always to be one level lower). So I ended up to
hack this by inserting the needed ### heading via Makefile for
the LaTeX version only (the HTML version would be messed up if
this header is present).

Drawbacks:
 * I use `git checkout` to make sure the `index.rst` is
   pristine before adding my the header. This means any
   changes to that file need to be committed before
   running `make`.
 * The PDF version contains the "Front Matter" entry in the TOC