Modified Makefile so that the S3 bucket (and path within that bucket) can be specified
The syntax is to run make with S3BUCKET=... on the command line.  Of course, you need to
have the proper credentials to write to the bucket as well.  I added this in preparation
for having an automated way for forks to publish their own versions of the book
automatically to a 'devs' site.