Original equations for line were both confusing and wrong (no coincidence)
This addresses issue #287.  Honestly, I don't know where I got the derivation for
the equations I was using.  They were unnecessarily complicated and definitely
wrong (at least in some cases).  Erik Vetter pointed this out and he is completely
correct.  I've completely redone the equation for the line and the Line and LineWithProtected
models to demonstrate what I wanted to about the language but without the stupidly
complex equations.  This is much clearer now.

This closes #287