Fixed two build issues
One is with Matplotlib and the chattering
example.  Apparently, that number of
data points gives it trouble.

The second fix involved OpenModelica
and making sure all results are written
out.  Apparently, the newer versions of
OM changed what was written out by
default.  This just instructs it to write
everything to the results file.  This probably
makes the builds take longer, but at least
the build succeeds.