Strip out inclusion of mathjax script tag
This tag wasn't doing anything, but no
sense in having the browser load this
if it doesn't need.

It gets included by the MathJax extension
in Sphinx.  If we remove that extension,
Sphinx complains that there is no
extension to handle inline math (which
would actually be fine), but I don't like
the warning.  So I keep the MathJax
extension in the build, let it add the
script tag, inline the math using
mathjax-node-page and then remove
the script tag.  That way, no warnings
show up.