Removing  labels that are not followed by a section
These will cause multiple labels being defined. If one wants to link to a lteral iunclude then this needs to be done via the `:name:` field but that *must* follow a `:caption:` field. Since I did not want to add additional captions and only one of the removed labels was referenced I simply pointedt the used reference to the section instead.