Using inheritance instead of modifications
It turns out that for some reason a run-time application of a modification to a case doesn't
work in this case because of some issue with the parameter being bound.  I checked and this is
the only case where I actually do this.  So rather than try to figure out why OM is binding this
parameter, I'll just make an additional model.