Cleans up the set of parameters shown in the browser
Lots of parameters were getting listed that really were not relevant because they didn't impact
the simulation.  These can be identified in the output of OpenModelica by looking for the
isValueChangeable field.  I'm not sure what a variable with parameter variability means if it
isn't changeable (probably bound, but recorded anyway).  But I'm removing them from the set of
parameters shown in the GUI.