Fixed one lingering issue with #310
The issue is that earlier discussion of the chemical example problem contains
references to the R variable which I eliminated in the component oriented presentation
of the problem.  So I made a slight change to make them consistent.  Since the R and
S species aren't actually in the model, it was just a minor change to the text.