Building on Mac M1
I had a hell of a time figuring this out.  Using x86-64 images on
an M1 Mac, I got all kinds of strange errors from clang.  The issue
was the way it detected the underlying CPU.  Previously, it used
the -march=native CLI flag and that got it in all kinds of trouble.
Because the book-builder Docker image is an x86-84 image, I've
opted to just force the build to use a generic x86-64 architecture
for all builds.  This seems safe and keeps clang out of trouble.