Supports building all results files in a single pass
Now only is this going to be more time efficient because it won't
load the MSL for each model, it will also help avoid the issue I've
been having with intermittend memory allocation issues on M1 Macs.

I was hoping combining all models would speed this up even more,
but it still takes about 30 minutes on my M1 to build all results.
But I haven't seen it fail in the middle like it was when I built
them individually.