chore: implementing a batch mode
The goal here is the 'smartly' run a bunch of cases.  My hope is to simplify the build process in
several ways.  First, no more Makefile to generate...the logic will be built into mobe.  In addition,
the results and executables should be stored in a specific directory (that we really don't need to
look into, since it contains so many files).  Finally, it should be smart enough to know when to
rebuild things.  The model hash can be used to determine if a recompile is necessary and dates on
results files could be used to determine whether the simulation needs to be rerun.